requirements.txt
pip=21.2.4
Jinja2=3.0.1 
arkupSafe=2.0.1 
Werkzeug=2.0.1
click=8.0.1 
colorama=0.4.4 
dataclasses=0.8 
flask=2.0.1
 importlib-metadata=4.8.1 
 itsdangerous=2.0.1 
 typing-extensions=3.10.0.2
  zipp=3.5.0 
 SQLAlchemy=1.4.25 
 flask-sqlalchemy=2.5.1 
 greenlet=1.1.1
  psycopg2=2.9.1
  PyJWT=2.1.0
  flask-swagger-ui=3.36.0
  flask-marshmallow=0.14.0
   marshmallow=3.13.0 
   six=1.16.0
   flask-restplus=0.13.0 
   jinja2=3.0.0rc2